ASoC: msm: qdsp6v2: Fix custom stereo state mismatch with DAP
Custom stereo state mismatch is seen when DAP on/bypass is
concurrently performed with custom stereo on/off. Store the
custom stereo state in ds2 module and configure the custom
stereo module appropriately. When DAP is turned on and custom
stereo is on turn off the custom stereo on QTI channel mixer
and turn on custom stereon on DAP. When DAP is bypassed and
custom stereo is on turn off custom stereo on DAP module and
turn on custom stereo on QTI module.

CRs-Fixed: 676884
Change-Id: Ice6cf2290a6e4de08068a0a63a3dd829386003fe
Signed-off-by: Krishnankutty Kolathappilly <kkolat@codeaurora.org>